requirements_bk.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.1
#  numpy==1.23.3
#  opencv-python==4.6.0.66
openpyxl==3.0.10
Pillow==9.2.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
